requirements.txt
flask
discord-interactions
supabase
requests
python-dotenv